requirements.txt
pyrogram
TgCrypto
py-tgcalls==0.5.5
python-dotenv
youtube_dl
youtube_search_python==1.4.6
aiofiles 
requests
aiohttp
asyncio
youtube_search
search_engine_parser
wget
yt_dlp
